Critical bug in iterate fixed
yea, so you're not supposed to call iterate... well, trial calls it a
lot, and good thing too... the logic of the main event loop was broken
because of the core timer control mechanism which uses a single shot
timer... looking at profile call data, the problem was glaring.
single line fix basically.

looking at the profile data also indicates that this mechanism is
unnecessary and is next on the list